reframing FACE vs BOUNDARY_FACE to compile-time, not runtime
